Move parsing logic into its own module.
For better sharing with done-worker-autorender
