+++ src/autorender.js
@@ -1,10 +1,9 @@
 define([
 	"@loader",
 	"module",
-	"can/view/stache/intermediate_and_imports",
+	"./parse",
 	"./template"
-], function(loader, module, getIntermediateAndImports, template){
-
+], function(loader, module, parse, template){
 	var main;
 
 	var isNode = typeof process === "object" &&
@@ -40,31 +39,13 @@ define([
 	}
 
 	function translate(load){
-		var intermediateAndImports = getIntermediateAndImports(load.source);
-
-		var ases = intermediateAndImports.ases;
-		var imports = intermediateAndImports.imports;
-		var args = [];
-		can.each(ases, function(from, name){
-			// Move the as to the front of the array.
-			imports.splice(imports.indexOf(from), 1);
-			imports.unshift(from);
-
-			// Get rid of @ character for the viewModel
-			if(name[0] === "@") {
-				name = name.substr(1);
-			}
-			args.unshift(name);
-		});
-		imports.unshift("can/view/stache/stache");
-		args.unshift("stache");
-
+		var result = parse(load.source);
 
 		var output = template({
-			imports: JSON.stringify(intermediateAndImports.imports),
-			args: args.join(", "),
-			intermediate: JSON.stringify(intermediateAndImports.intermediate),
-			ases: can.map(ases, function(from, name){
+			imports: JSON.stringify(result.imports),
+			args: result.args.join(", "),
+			intermediate: JSON.stringify(result.intermediate),
+			ases: can.map(result.ases, function(from, name){
 				return "\t" + name + ": " + name +"['default'] || " + name;
 			}).join(",\n")
 		});
